+++ .gitignore
@@ -22,6 +22,7 @@ resources
 .pytest_cache
 /expressions.tokens
 tags
+!db/tags/
 testlog
 test/*/*.reject
 .vscode
