+++ alternator/executor.hh
@@ -81,10 +81,9 @@ namespace parsed {
 class path;
 };
 
-const std::map<sstring, sstring>& get_tags_of_table(schema_ptr schema);
-std::optional<std::string> find_tag(const schema& s, const sstring& tag);
-future<> update_tags(service::migration_manager& mm, schema_ptr schema, std::map<sstring, sstring>&& tags_map);
 schema_ptr get_table(service::storage_proxy& proxy, const rjson::value& request);
+// Wraps the db::get_tags_of_table and throws if the table is missing the tags extension.
+const std::map<sstring, sstring>& get_tags_of_table_or_throw(schema_ptr schema);
 
 // An attribute_path_map object is used to hold data for various attributes
 // paths (parsed::path) in a hierarchy of attribute paths. Each attribute path
