+++ configure.py
@@ -881,6 +881,7 @@ scylla_core = (['replica/database.cc',
                 'db/large_data_handler.cc',
                 'db/marshal/type_parser.cc',
                 'db/batchlog_manager.cc',
+                'db/tags/utils.cc',
                 'db/view/view.cc',
                 'db/view/view_update_generator.cc',
                 'db/view/row_locking.cc',
